pyproject.toml
[project]
name = "mcp-file-server"
version = "1.0.0"
description = "MCP File System Server with Docker and ngrok"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []

